refactor(client): renames auth context folder
