+++ client/src/app.tsx
@@ -10,7 +10,7 @@ import {
 	SIGN_IN_PATH,
 	SIGN_UP_PATH
 } from 'constants/paths';
-import AuthContext from 'context/profile/auth.context';
+import AuthContext from 'context/auth/auth.context';
 import { Account } from 'pages/account';
 import { Code } from 'pages/code';
 import { CreateProject } from 'pages/create-project';
