+++ client/src/context/auth/auth.context.tsx
@@ -1,7 +1,7 @@
 import { createContext, ReactNode, useEffect, useState } from 'react';
 import { ProfileQuery } from 'graphql/__generated__/graphql';
 import { ApolloQueryResult, OperationVariables } from '@apollo/client';
-import { useProfile } from 'context/profile/profile.service';
+import { useProfile } from 'context/auth/profile.service';
 export interface AuthContextProps {
 	profile: ProfileQuery | null;
 	isLoading: boolean;
